refactor(ensure): port max-line-length and its tests to ts
